requirements.txt
google-cloud-texttospeech
google-cloud-monitoring
google-auth
python-dateutil
requests
beautifulsoup4
tqdm
pydub
